requirements.txt
python-multipart
python-dotenv
openai
uvicorn
fastapi